fixed type error with message

+++ suripu-core/src/main/java/com/hello/suripu/core/models/Insights/Message/WakeVarianceMsgEN.java
@@ -5,19 +5,19 @@ package com.hello.suripu.core.models.Insights.Message;
  */
 
 public class WakeVarianceMsgEN {
-    public static Text getWakeVarianceLow(final double wakeVariance, final int percentile) {
+    public static Text getWakeVarianceLow(final int wakeVariance, final int percentile) {
         return new Text("Hello, very regular",
                 String.format("You have very consistent sleep habits. Your wake time variance of %d is low, ", wakeVariance) +
                         String.format("it is **lower than** %d%% of all Sense users.", 100 - percentile));
     }
 
-    public static Text getWakeVarianceNotLowEnough(final double wakeVariance, final int percentile) {
+    public static Text getWakeVarianceNotLowEnough(final int wakeVariance, final int percentile) {
         return new Text("Hello, regular",
                 String.format("You have consistent sleep habits. Your wake time variance of %d is low, ", wakeVariance) +
                         String.format("it is **lower than** %d%% of all Sense users.", 100 - percentile));
     }
 
-    public static Text getWakeVarianceHigh(final double wakeVariance, final int percentile) {
+    public static Text getWakeVarianceHigh(final int wakeVariance, final int percentile) {
         return new Text("Hello, irregular",
                 String.format("Your sleep habits are a bit irregular. Your wake time variance of %d is a bit high, ", wakeVariance) +
                         String.format("it is **higher than** %d%% of all Sense users.", percentile) +
@@ -25,7 +25,7 @@ public class WakeVarianceMsgEN {
 
     }
 
-    public static Text getWakeVarianceTooHigh(final double wakeVariance, final int percentile) {
+    public static Text getWakeVarianceTooHigh(final int wakeVariance, final int percentile) {
         return new Text("Hello, very irregular",
                 String.format("Your sleep habits are irregular. Your wake time variance of %d is high", wakeVariance) +
                         String.format("it is **higher than** %d%% of all Sense users.", percentile) +
